+++ src/components/planner/PlannerView.js
@@ -1,53 +1,126 @@
+import { Close, Delete } from "@mui/icons-material";
+import dayjs from "dayjs";
 import { useState } from "react";
 import { useRecoilState } from "recoil";
 
-const PlannerView = () => {
-  //마커 오버레이 여닫음
-  const [openOverlay, setOpenOverlay] = useState(null);
-  //플래너 창 여닫음
-  const [openPlanner, setOpenPlanner] = useState(false);
-  //"플래너에 추가하기" 창 여닫음
-  const [openPlanningModal, setOpenPlanningModal] = useState(null);
-  //플래너에 추가한 장소 리스트
-  const [plannedPlaceList, setPlannedPlaceList] = useState([]);
-  //플래너 제목
-  const [planName, setPlanName] = useState("");
-  //플래너 저장 창 여닫음
-  const [openSaveModal, setOpenSaveModal] = useState(false);
-  //장소 리스트(임시 데이터)
-  const [placeList, setPlaceList] = useState([]);
-  //정렬 옵션(1:거리순, 2:리뷰많은순, 3:이름순)
-  const [sortOption, setSortOption] = useState(1);
-  //필터 옵션(null:전체, 1:숙박시설, 2:음식점, 3:그외)
-  const [filterOption, setFilterOption] = useState(null);
-  //유저닉네임
-  const [loginNickname, setLoginNickname] = useRecoilState(loginNicknameState);
-  //플래너 소유자(작성자) 여부
-  const [isOwner, setIsOwner] = useState(true);
+const PlannerView = (props) => {
+  const {
+    openPlanner,
+    setOpenPlanner,
+    plannedPlaceList,
+    setPlannedPlaceList,
+    planName,
+    setPlanName,
+    plannerMode,
+  } = props;
 
-  //↓ 맵 관련 STATE
-  //현재 보이는 지도 화면
-  const [mapBounds, setMapBounds] = useState(null);
-  //지도 중심좌표(화면 이동 시 사용)
-  const [mapCenter, setMapCenter] = useState({
-    lat: 37.5341338,
-    lng: 126.897333254,
-  });
-  //유저가 클릭한 지도 위치
-  const [userMarker, setUserMarker] = useState(null);
-  //유저 클릭 위치를 중심으로 하는 반경 범위
-  const [userRadius, setUserRadius] = useState(1000);
-
-  //↓ 플래너 상태 판별용 STATE
-  //주소에서 받은 planNo
-  const { planNo } = useParams();
-  //플래너 모드: view, write, null
-  const [plannerMode, setPlannerMode] = useState(null);
-
+  const handleDeletePlace = (odr) => {
+    const newList = plannedPlaceList.filter((item) => item.order !== odr);
+    if (odr === 0 && newList.length > 0) {
+      newList[0].transport = "";
+    }
+    newList.forEach((item, i) => {
+      item.order = i;
+    });
+    setPlannedPlaceList(newList);
+  };
   return (
     <>
-      <div></div>
+      {openPlanner ? (
+        <Planner
+          setOpenPlanner={setOpenPlanner}
+          plannedPlaceList={plannedPlaceList}
+          handleDeletePlace={handleDeletePlace}
+          planName={planName}
+          setPlanName={setPlanName}
+          plannerMode={plannerMode}
+        />
+      ) : (
+        <div
+          className={`planner-close-btn ${
+            plannerMode === "view" ? "full" : ""
+          }`}
+          onClick={() => setOpenPlanner(true)}
+        >
+          <span>📆</span>
+          <p>플래너</p>
+        </div>
+      )}
     </>
+  );
+};
+
+// 여행 플래너 출력 창
+const Planner = (props) => {
+  const setOpenPlanner = props.setOpenPlanner;
+  const plannedPlaceList = props.plannedPlaceList;
+  const handleDeletePlace = props.handleDeletePlace;
+  const [planName, setPlanName] = [props.planName, props.setPlanName];
+  const plannerMode = props.plannerMode;
+
+  return (
+    <div className={`planner-wrap ${plannerMode === "view" ? "full" : ""}`}>
+      <Close className="close-btn" onClick={() => setOpenPlanner(false)} />
+      {plannedPlaceList.length === 0 && (
+        <div className="empty-plan">플래너가 비어 있습니다...</div>
+      )}
+      <div className="plan-name">
+        <input
+          type="text"
+          placeholder="플래너 이름을 작성하세요"
+          value={planName}
+          onChange={(e) => setPlanName(e.target.value)}
+          readOnly={plannerMode === "view"}
+        />
+      </div>
+      {plannedPlaceList
+        .sort((a, b) => a.order - b.order)
+        .map((p, idx) => {
+          const isDateChanged =
+            idx === 0 ||
+            p.itineraryDate !== plannedPlaceList[idx - 1].itineraryDate;
+          const showTransport = idx !== 0;
+          if (p.placeThumb == null) {
+            p.placeThumb = "/image/place_default_img.png";
+          }
+          return (
+            <div key={"planned-" + idx}>
+              {showTransport && (
+                <div className="planned-transport">
+                  <span>↓</span>
+                  <span>{p.transport}(으)로 이동</span>
+                </div>
+              )}
+              {isDateChanged && (
+                <div className="planned-date">
+                  ㅡ {dayjs(p.itineraryDate).format("YYYY년 M월 D일")} ㅡ
+                </div>
+              )}
+              <div className="planned-item">
+                <img
+                  className="planned-img"
+                  src={p.placeThumb}
+                  alt="테스트"
+                  width="50px"
+                  height="50px"
+                />
+                <div className="place-item">
+                  <div className="place-title-wrap">
+                    <span className="place-title">{p.placeTitle}</span>
+                    <span className="place-type">{p.placeType}</span>
+                  </div>
+                  <div className="place-addr">{p.placeAddr}</div>
+                </div>
+                {plannerMode === "write" && (
+                  <div className="planner-del-btn">
+                    <Delete onClick={() => handleDeletePlace(p.order)} />
+                  </div>
+                )}
+              </div>
+            </div>
+          );
+        })}
+    </div>
   );
 };
